requirements.txt
amqp==5.1.1
APScheduler==3.10.4
asgiref==3.7.2
astroid==3.0.1
attrs==23.1.0
billiard==4.1.0
cattrs==23.1.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.7.0
cryptography==41.0.4
defusedxml==0.8.0rc2
dill==0.3.7
Django~=4.2.6
django-allauth==0.57.0
django-apscheduler==0.6.2
django-filter==23.3
idna==3.4
isort==5.12.0
kombu==5.3.2
lazy-object-proxy==1.9.0
lsprotocol==2023.0.0b1
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
pycparser==2.21
pygls==1.1.1
PyJWT==2.8.0
pylint==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.1.1
ruff-lsp==0.0.41
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.1
typeguard==3.0.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
wrapt==1.16.0rc1
django-modeltranslation~=0.18.11